1_create_dataset/requirements.txt
pydantic-ai-slim[openrouter]
pydantic
tqdm
python-slugify
python-dotenv
ruff
